Move remaining functions in opoutcome_utils.py to utils.py
